fix: more instrumentor example to top level
