--- sdks/python/examples/opentelemetry_instrumentation/client.py
@@ -1,3 +1,0 @@
-from hatchet_sdk import Hatchet
-
-hatchet = Hatchet(debug=True)
